Move MuJoCo-related files to simulation/multiBodyDynamics
