+++ src/simulation/multiBodyDynamics/_GeneralModuleFiles/MJEquality.cpp
