+++ src/simulation/multiBodyDynamics/_GeneralModuleFiles/MJEquality.h
