+++ src/simulation/multiBodyDynamics/_GeneralModuleFiles/MJEquality.swg
