+++ src/simulation/multiBodyDynamics/_GeneralModuleFiles/MJUtils.cpp
