+++ src/simulation/multiBodyDynamics/_GeneralModuleFiles/_UnitTest/test_sat_2_changed_mass.xml
